Fix memory leaks for array_merge
This commit fixes the memory leaks present in array_merge_test.cpp. The leaks were coming from the fact that some memory is allocated during array_merge, but it is the responsibility of the function that calls array_merge to free this memory. The tests were not doing that, so we added free(result) statements to free this memory. Additionally, there were a couple tests where memory was allocated for a[i] elements, and this memory was never freed so  we added a for loop that frees these elements.

Co-authored-by: Joshua Eklund <eklun124@umn.edu>